requirements.txt
PyYAML==3.11
Unidecode==0.04.19
beautifulsoup4==4.4.1
markovify==0.2.5
pyaml==15.8.2
python-wordpress-xmlrpc==2.3
requests==2.9.1
